pyproject.toml
[tool.poetry]
name = "tass"
version = "0.1.0"
description = ""
authors = ["Milosz Lopatto <milosz.lopatto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.2"
fpdf2 = "^2.8.2"
pymongo = "^4.10.1"
matplotlib = "^3.10.0"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
